ML/Fraud/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
xgboost>=1.5.0
easyocr>=1.6.0
pytesseract>=0.3.8
pdfminer.six>=20211012
fastapi>=0.68.0
uvicorn>=0.15.0
joblib>=1.1.0
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
transformers>=4.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
pillow>=8.3.0
numpy>=1.21.0
pandas>=1.3.0
